refactor: GameScore객체 분리
